BC may need to replace the animated logotext
