+++ core/components/controls/Features/components/Wall.js
@@ -1,8 +1,5 @@
 import styled, { css } from "styled-components";
 
-import { LabelWrap } from "../../Docket";
-import { c_black, c_white } from "../../../../../constants/styles/colors";
-
 export default styled.div`
   /* this allows better position for scrollbars */
   height: 8em;
@@ -25,35 +22,5 @@ export const unstyledLinks = css`
   a:active,
   a:focus {
     background: 0 0;
-  }
-`;
-export const CollectionDescription = styled.blockquote`
-  margin: 1.5em auto !important;
-  ${unstyledLinks};
-
-  ${props =>
-    !props.showDescription &&
-    `
-      &::before, &::after { content: "" !important; };
-      padding: 1.15em !important;
-      border-top: none !important;
-      margin: calc(1.5em + 8px) auto .5em !important;
-    `}
-`;
-
-export const BreadcrumbsWrap = styled(LabelWrap)`
-  top: 0;
-  font-style: normal;
-  width: auto;
-  height: 2em;
-  ${unstyledLinks};
-  label {
-    line-height: 1.25em;
-  }
-  a:last-child {
-    label {
-      background: ${c_black};
-      color: ${c_white};
-    }
   }
 `;
